+++ sqlalchemy_unchained/base_model_metaclass.py
@@ -122,10 +122,10 @@ class BindMetaMixin:
 
 class DeclarativeMeta(NameMetaMixin, BindMetaMixin, BaseDeclarativeMeta):
     def __new__(mcs, name, bases, clsdict):
-        from .model_registry import ModelRegistry
+        from .model_registry import _ModelRegistry
 
         mcs_args = McsArgs(mcs, name, bases, clsdict)
-        ModelRegistry()._ensure_correct_base_model(mcs_args)
+        _ModelRegistry()._ensure_correct_base_model(mcs_args)
         options_factory = apply_factory_meta_options(
             mcs_args, default_factory_class=ModelMetaOptionsFactory)
 
@@ -158,7 +158,7 @@ class DeclarativeMeta(NameMetaMixin, BindMetaMixin, BaseDeclarativeMeta):
                     validators[column].append(attr_name)
         clsdict['__validators__'] = validators
 
-        ModelRegistry().register_new(mcs_args)
+        _ModelRegistry().register_new(mcs_args)
         return super().__new__(*mcs_args)
 
     def __init__(cls, name, bases, clsdict):
@@ -182,15 +182,15 @@ class DeclarativeMeta(NameMetaMixin, BindMetaMixin, BaseDeclarativeMeta):
         if should_set_tablename(cls):
             cls.__tablename__ = snake_case(cls.__name__)
 
-        from .model_registry import ModelRegistry
-        if not ModelRegistry().enable_lazy_mapping or \
+        from .model_registry import _ModelRegistry
+        if not _ModelRegistry().enable_lazy_mapping or \
                 (not cls.Meta.abstract and not cls.Meta.lazy_mapped):
             cls._pre_mcs_init()
             super().__init__(name, bases, clsdict)
             cls._post_mcs_init()
 
         if not cls.Meta.abstract:
-            ModelRegistry().register(McsInitArgs(cls, name, bases, clsdict))
+            _ModelRegistry().register(McsInitArgs(cls, name, bases, clsdict))
 
     def _pre_mcs_init(cls):
         """